Fix type conversion and namespace errors across multiple files
- EventFactory.cs: Added EventType overload for MapCategoryToIssue method
- GameManagerIntegration.cs: Fixed LogPurchase call - cast to long instead of float
- PlayerResponseSystem.cs: Fixed ResponseHistory.Add to use string instead of PlayerResponse
- NewsSystemOrchestrator.cs: Fixed ResponseResult namespace conflict by using explicit
  Consequences.ResponseResult return type and converting for event args

+++ Assets/Scripts/Core/GameManagerIntegration.cs
@@ -697,7 +697,7 @@ namespace ElectionEmpire.Core
         {
             if (purchase != null && !purchase.IsRefunded)
             {
-                Analytics.LogPurchase(purchase.PackageId, purchase.Currency, (float)purchase.AmountPaid, purchase.IsVerified);
+                Analytics.LogPurchase(purchase.PackageId, purchase.Currency, (long)purchase.AmountPaid, purchase.IsVerified);
             }
         }
